refactor: mock system chats
